+++ RELEASES.md
@@ -94,6 +94,7 @@ Upcoming
 - Sort both tables in "/rest/webservices" on name
 - Bugfix layout "Security Role Bindings" in console function "Show Security Items"
 - Add lookupDestination attribute to JMSFacade to have possibility to connect to queues directly instead of via JNDI
+- Add ability to log info from incoming messages on EsbJmsListener extracted by an xPath